+++ src/main/resources/mybatis/mappers/ReportFileMapper.xml
@@ -55,7 +55,7 @@
             original_file_name,
             stored_file_name,
             file_size,
-            upload_date,
+            uploaded_at,
             uploaded_by,
             file_id_usage
         ) VALUES (
@@ -66,7 +66,7 @@
                      #{originalFileName},
                      #{storedFileName},
                      #{fileSize},
-                     #{uploadDate},
+                     #{uploadedAt},
                      #{uploadedBy},
                      #{fileIdUsage}
                  )
@@ -81,7 +81,7 @@
             original_file_name,
             stored_file_name,
             file_size,
-            upload_date,
+            uploaded_at,
             uploaded_by,
             file_id_usage
         ) VALUES (
@@ -92,7 +92,7 @@
             #{originalFileName},
             #{storedFileName},
             #{fileSize},
-            #{uploadDate},
+            #{uploadedAt},
             #{uploadedBy},
             #{fileIdUsage}
             )
